+++ common/src/main/java/net/spell_engine/mixin/client/ClientPlayerEntityMixin.java
@@ -4,15 +4,14 @@ import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.network.ClientPlayNetworkHandler;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
-import net.spell_engine.Platform;
 import net.spell_engine.SpellEngineMod;
 import net.spell_engine.api.effect.EntityActionsAllowed;
 import net.spell_engine.api.spell.Spell;
+import net.spell_engine.PlatformClient;
 import net.spell_engine.client.SpellEngineClient;
 import net.spell_engine.client.input.SpellHotbar;
 import net.spell_engine.internals.SpellHelper;
@@ -249,7 +248,7 @@ public abstract class ClientPlayerEntityMixin implements SpellCasterClient {
         updateSpellCast();
         var player = player();
         if (isBeaming()) {
-            Platform.util().sendVanillaPacket_C2S(player, new PlayerMoveC2SPacket.Full(
+            PlatformClient.util().sendVanillaPacket_C2S(player, new PlayerMoveC2SPacket.Full(
                     player.getX(), player.getY(), player.getZ(),
                     player.getYaw(), player.getPitch(),
                     player.isOnGround())