+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.8.2
+
+Functional changes:
+- Fix crash on archery bow-melee swap
+
 # 1.8.1
 
 Functional changes: